backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
trafilatura==1.12.2
cohere==5.11.4
sse-starlette==2.1.3
